Add form for editing source
This is one form that changes both the Source entity (representing the
pack as a whole) and the Release entity (with information about the most
recent release). I am trying to avoid having a separate page for
releases, because it is just extra UI load for users and will 99% of the
time not be necessary—and the remaining cases can be done through appeal
to the admin.